@@ -6,14 +6,14 @@ public static class SourceGenerationHelper
     {
         var r = 
             $$"""
-              using System;
+              // <auto-generated />
+              #nullable enable
 
-              namespace Hubs.InterfaceGenerators
+              namespace Hubs.InterfaceGenerators;
+
+              public class {{value.GeneratedClassName}}Impl : {{value.InterfaceName}}
               {
-                  public class {{value.GeneratedClassName}}Impl : {{value.InterfaceName}}
-                  {
               {{string.Join("\n\n", MethodsImplementation(value))}}
-                  }
               }
               """;
 
@@ -24,12 +24,12 @@ public static class SourceGenerationHelper
             foreach (var signature in value.MethodSignatures)
             {
                 yield return
-                    $$"""
-                              async {{signature}} 
-                              {
-                                  throw new NotImplementedException();
-                              }
-                      """;
+                $$"""
+                      async {{signature}} 
+                      {
+                          throw new NotImplementedException();
+                      }
+                  """;
             }
         }
     }
