@@ -1,23 +1,23 @@
 ﻿//HintName: Contract1.g.cs
-using System;
+// <auto-generated />
+#nullable enable
 
-namespace Hubs.InterfaceGenerators
+namespace Hubs.InterfaceGenerators;
+
+public class Contract1Impl : global::IContract1
 {
-    public class Contract1Impl : IContract1
+    async global::System.Threading.Tasks.Task global::IContract1.Method1() 
     {
-        async System.Threading.Tasks.Task IContract1.Method1() 
-        {
-            throw new NotImplementedException();
-        }
+        throw new NotImplementedException();
+    }
 
-        async System.Threading.Tasks.Task IContract1.Method2(int i) 
-        {
-            throw new NotImplementedException();
-        }
+    async global::System.Threading.Tasks.Task global::IContract1.Method2(int i) 
+    {
+        throw new NotImplementedException();
+    }
 
-        async System.Threading.Tasks.Task IContract1.Method3(object i) 
-        {
-            throw new NotImplementedException();
-        }
+    async global::System.Threading.Tasks.Task global::IContract1.Method3(object i) 
+    {
+        throw new NotImplementedException();
     }
 }
