@@ -1,38 +1,38 @@
 ﻿//HintName: Contract2.g.cs
-using System;
+// <auto-generated />
+#nullable enable
 
-namespace Hubs.InterfaceGenerators
+namespace Hubs.InterfaceGenerators;
+
+public class Contract2Impl : global::IContract2
 {
-    public class Contract2Impl : IContract2
+    async global::System.Threading.Tasks.Task global::IContract1.Method1() 
     {
-        async System.Threading.Tasks.Task IContract1.Method1() 
-        {
-            throw new NotImplementedException();
-        }
+        throw new NotImplementedException();
+    }
 
-        async System.Threading.Tasks.Task IContract1.Method2(int i) 
-        {
-            throw new NotImplementedException();
-        }
+    async global::System.Threading.Tasks.Task global::IContract1.Method2(int i) 
+    {
+        throw new NotImplementedException();
+    }
 
-        async System.Threading.Tasks.Task IContract1.Method3(object i) 
-        {
-            throw new NotImplementedException();
-        }
+    async global::System.Threading.Tasks.Task global::IContract1.Method3(object i) 
+    {
+        throw new NotImplementedException();
+    }
 
-        async System.Threading.Tasks.Task<int> IContract2.Method4() 
-        {
-            throw new NotImplementedException();
-        }
+    async global::System.Threading.Tasks.Task<int> global::IContract2.Method4() 
+    {
+        throw new NotImplementedException();
+    }
 
-        async System.Threading.Tasks.Task<int> IContract2.Method5(int i) 
-        {
-            throw new NotImplementedException();
-        }
+    async global::System.Threading.Tasks.Task<int> global::IContract2.Method5(int i) 
+    {
+        throw new NotImplementedException();
+    }
 
-        async System.Threading.Tasks.Task<int> IContract2.Method6(object i) 
-        {
-            throw new NotImplementedException();
-        }
+    async global::System.Threading.Tasks.Task<int> global::IContract2.Method6(object i) 
+    {
+        throw new NotImplementedException();
     }
 }
