@@ -1,38 +1,38 @@
 ﻿//HintName: Contract.g.cs
-using System;
+// <auto-generated />
+#nullable enable
 
-namespace Hubs.InterfaceGenerators
+namespace Hubs.InterfaceGenerators;
+
+public class ContractImpl : global::TestNamespace.IContract
 {
-    public class ContractImpl : TestNamespace.IContract
+    async global::System.Threading.Tasks.Task global::TestNamespace.IContract.Method1() 
     {
-        async System.Threading.Tasks.Task TestNamespace.IContract.Method1() 
-        {
-            throw new NotImplementedException();
-        }
+        throw new NotImplementedException();
+    }
 
-        async System.Threading.Tasks.Task TestNamespace.IContract.Method2(int i) 
-        {
-            throw new NotImplementedException();
-        }
+    async global::System.Threading.Tasks.Task global::TestNamespace.IContract.Method2(int i) 
+    {
+        throw new NotImplementedException();
+    }
 
-        async System.Threading.Tasks.Task TestNamespace.IContract.Method3(object i) 
-        {
-            throw new NotImplementedException();
-        }
+    async global::System.Threading.Tasks.Task global::TestNamespace.IContract.Method3(object i) 
+    {
+        throw new NotImplementedException();
+    }
 
-        async System.Threading.Tasks.Task<int> TestNamespace.IContract.Method4() 
-        {
-            throw new NotImplementedException();
-        }
+    async global::System.Threading.Tasks.Task<int> global::TestNamespace.IContract.Method4() 
+    {
+        throw new NotImplementedException();
+    }
 
-        async System.Threading.Tasks.Task<int> TestNamespace.IContract.Method5(int i) 
-        {
-            throw new NotImplementedException();
-        }
+    async global::System.Threading.Tasks.Task<int> global::TestNamespace.IContract.Method5(int i) 
+    {
+        throw new NotImplementedException();
+    }
 
-        async System.Threading.Tasks.Task<int> TestNamespace.IContract.Method6(object i) 
-        {
-            throw new NotImplementedException();
-        }
+    async global::System.Threading.Tasks.Task<int> global::TestNamespace.IContract.Method6(object i) 
+    {
+        throw new NotImplementedException();
     }
 }
