@@ -1,38 +1,38 @@
 ﻿//HintName: Contract.g.cs
-using System;
+// <auto-generated />
+#nullable enable
 
-namespace Hubs.InterfaceGenerators
+namespace Hubs.InterfaceGenerators;
+
+public class ContractImpl : global::IContract
 {
-    public class ContractImpl : IContract
+    async global::System.Threading.Tasks.Task global::IContract.Method1() 
     {
-        async System.Threading.Tasks.Task IContract.Method1() 
-        {
-            throw new NotImplementedException();
-        }
+        throw new NotImplementedException();
+    }
 
-        async System.Threading.Tasks.Task IContract.Method2(int i) 
-        {
-            throw new NotImplementedException();
-        }
+    async global::System.Threading.Tasks.Task global::IContract.Method2(int i) 
+    {
+        throw new NotImplementedException();
+    }
 
-        async System.Threading.Tasks.Task IContract.Method3(object i) 
-        {
-            throw new NotImplementedException();
-        }
+    async global::System.Threading.Tasks.Task global::IContract.Method3(object i) 
+    {
+        throw new NotImplementedException();
+    }
 
-        async System.Threading.Tasks.Task<int> IContract.Method4() 
-        {
-            throw new NotImplementedException();
-        }
+    async global::System.Threading.Tasks.Task<int> global::IContract.Method4() 
+    {
+        throw new NotImplementedException();
+    }
 
-        async System.Threading.Tasks.Task<int> IContract.Method5(int i) 
-        {
-            throw new NotImplementedException();
-        }
+    async global::System.Threading.Tasks.Task<int> global::IContract.Method5(int i) 
+    {
+        throw new NotImplementedException();
+    }
 
-        async System.Threading.Tasks.Task<int> IContract.Method6(object i) 
-        {
-            throw new NotImplementedException();
-        }
+    async global::System.Threading.Tasks.Task<int> global::IContract.Method6(object i) 
+    {
+        throw new NotImplementedException();
     }
 }
